Move main reducer to index.js
